Clean up other CMR files (SearchBar, AddNewMenu)
